rename files with dash case, and return dispatch or promise.resolve in the promise.all map when deleting an offering
